#57 HTML構造を綺麗に修正しリファクタした。 - li li となっていた箇所をul liの順に修正 - 抽象度が高く役割が不明確だった関数およびコンポーネントを適切な名前に訂正 - Hydrationのみdev環境が原因か解消できなかった。
