+++ src/components/btn/TatoeListAddNewTatoeBtn.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
 
-export const TatoeListAddTatoeBtn = () => {
+export const TatoeListAddNewTatoeBtn = () => {
   return (
     <div
       className='
